Revert extraction
